+++ example/lib/main.dart
@@ -48,17 +48,32 @@ class _MyAppState extends State<MyApp> {
         setState(() {
           debugLable = "flutter onReceiveNotificationAuthorization: $message";
         });
-      },onNotifyMessageUnShow:
-          (Map<String, dynamic> message) async {
+      }, onNotifyMessageUnShow: (Map<String, dynamic> message) async {
         print("flutter onNotifyMessageUnShow: $message");
         setState(() {
           debugLable = "flutter onNotifyMessageUnShow: $message";
+        });
+      }, onInAppMessageShow: (Map<String, dynamic> message) async {
+        print("flutter onInAppMessageShow: $message");
+        setState(() {
+          debugLable = "flutter onInAppMessageShow: $message";
+        });
+      }, onInAppMessageClick: (Map<String, dynamic> message) async {
+        print("flutter onInAppMessageClick: $message");
+        setState(() {
+          debugLable = "flutter onInAppMessageClick: $message";
+        });
+      }, onConnected: (Map<String, dynamic> message) async {
+        print("flutter onConnected: $message");
+        setState(() {
+          debugLable = "flutter onConnected: $message";
         });
       });
     } on PlatformException {
       platformVersion = 'Failed to get platform version.';
     }
 
+    jpush.setAuth(enable: true);
     jpush.setup(
       appKey: "xxxxx", //你自己应用的 AppKey
       channel: "theChannel",