+++ lib/jpush_flutter.dart
@@ -26,6 +26,9 @@ class JPush {
   EventHandler? _onReceiveMessage;
   EventHandler? _onReceiveNotificationAuthorization;
   EventHandler? _onNotifyMessageUnShow;
+  EventHandler? _onConnected;
+  EventHandler? _onInAppMessageClick;
+  EventHandler? _onInAppMessageShow;
   void setup({
     String appKey = '',
     bool production = false,
@@ -43,15 +46,18 @@ class JPush {
   }
 
   //APP活跃在前台时是否展示通知
-  void setUnShowAtTheForeground({
-    bool unShow = false,
-  }) {
+  void setUnShowAtTheForeground({bool unShow = false}) {
     print(flutter_log + "setUnShowAtTheForeground:");
     _channel.invokeMethod('setUnShowAtTheForeground', {'UnShow': unShow});
   }
 
   void setWakeEnable({bool enable = false}) {
     _channel.invokeMethod('setWakeEnable', {'enable': enable});
+  }
+
+  void setAuth({bool enable = true}) {
+    print(flutter_log + "setAuth:");
+    _channel.invokeMethod('setAuth', {'enable': enable});
   }
 
   ///
@@ -63,6 +69,9 @@ class JPush {
     EventHandler? onReceiveMessage,
     EventHandler? onReceiveNotificationAuthorization,
     EventHandler? onNotifyMessageUnShow,
+    EventHandler? onConnected,
+    EventHandler? onInAppMessageClick,
+    EventHandler? onInAppMessageShow,
   }) {
     print(flutter_log + "addEventHandler:");
 
@@ -71,6 +80,9 @@ class JPush {
     _onReceiveMessage = onReceiveMessage;
     _onReceiveNotificationAuthorization = onReceiveNotificationAuthorization;
     _onNotifyMessageUnShow = onNotifyMessageUnShow;
+    _onConnected = onConnected;
+    _onInAppMessageClick = onInAppMessageClick;
+    _onInAppMessageShow = onInAppMessageShow;
     _channel.setMethodCallHandler(_handleMethod);
   }
 
@@ -89,6 +101,12 @@ class JPush {
             call.arguments.cast<String, dynamic>());
       case "onNotifyMessageUnShow":
         return _onNotifyMessageUnShow!(call.arguments.cast<String, dynamic>());
+      case "onConnected":
+        return _onConnected!(call.arguments.cast<String, dynamic>());
+      case "onInAppMessageClick":
+        return _onInAppMessageClick!(call.arguments.cast<String, dynamic>());
+      case "onInAppMessageShow":
+        return _onInAppMessageShow!(call.arguments.cast<String, dynamic>());
       default:
         throw new UnsupportedError("Unrecognized Event");
     }
@@ -182,18 +200,20 @@ class JPush {
         await _channel.invokeMethod('getAllTags');
     return result;
   }
+
   ///
   /// 获取所有当前绑定的 alias
   ///
-  /// @param {Function} success = ({"tags":[String]}) => {  }
+  /// @param {Function} success = ({"alias":String}) => {  }
   /// @param {Function} fail = ({"errorCode":int}) => {  }
   ///
   Future<Map<dynamic, dynamic>> getAlias() async {
     print(flutter_log + "getAlias:");
     final Map<dynamic, dynamic> result =
-    await _channel.invokeMethod('getAlias');
+        await _channel.invokeMethod('getAlias');
     return result;
   }
+
   ///
   /// 重置 alias.
   ///
@@ -208,6 +228,11 @@ class JPush {
     final Map<dynamic, dynamic> result =
         await _channel.invokeMethod('setAlias', alias);
     return result;
+  }
+
+  void testCountryCode(String code) {
+    print(flutter_log + "testCountryCode:" + code);
+    _channel.invokeMethod('testCountryCode', code);
   }
 
   ///